RFC: Remove zero initialization on Integrator, Derivative, and FirstOrder
As started to be discussed in the other thread, these are the only three `= 0` in this package that seem like they have a reasonable argument for it. There is a use case for which the integrator's starting value could be defined by an outside value, in which case this no `= 0` default is better as you will otherwise "accidentally" add an extra initial equation and fail initialization, unless you do `= nothing` (needs to be added).

But, is this better for the normal user?